events-api/requirements.txt
flask==3.0.2
redis==5.0.4
pika==1.3.2
python-dotenv==1.0.1
